% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_Rgression_HW_Package.R
\name{summarise_group_data}
\alias{summarise_group_data}
\title{Get summary and regression of data frame}
\usage{
summarise_group_data(data, group_id, obs, fun, ...)
}
\arguments{
\item{data}{A Data Frame}

\item{group_id}{Group Identification of data frame}

\item{obs}{Observations inside the data frame}

\item{fun}{Type of function}

\item{...}{Additional arguments inside function}
}
\value{
A data frame in the desired group and/or observation
}
\description{
Get summary and regression of data frame
}
\examples{
summarise_group_data <- function(data = iris.....)
}
